+++ lib/executable-config/lookup/src/Obelisk/Configs.hs
@@ -20,14 +20,17 @@ module Obelisk.Configs
 
 import Control.Applicative
 import Control.Monad
+import Control.Monad.Catch
 import Control.Monad.Base
 import Control.Monad.Fix
+import Control.Monad.Morph
 import Control.Monad.Primitive
 import Control.Monad.Ref
 import Control.Monad.IO.Class
-import Control.Monad.Trans.Class
 import Control.Monad.Trans.Control
 import Control.Monad.Trans.Reader
+import Control.Monad.Trans.State
+import qualified Control.Monad.Trans.State.Strict as Strict
 import Data.ByteString (ByteString)
 import Data.Map (Map)
 import qualified Data.Map as Map
@@ -62,6 +65,8 @@ instance HasConfigs m => HasConfigs (PostBuildT t m)
 instance HasConfigs m => HasConfigs (QueryT t q m)
 instance HasConfigs m => HasConfigs (ReaderT r m)
 instance HasConfigs m => HasConfigs (RequesterT t request response m)
+instance HasConfigs m => HasConfigs (StateT w m)
+instance HasConfigs m => HasConfigs (Strict.StateT w m)
 instance HasConfigs m => HasConfigs (StaticDomBuilderT t m)
 instance HasConfigs m => HasConfigs (TriggerEventT t m)
 
@@ -73,11 +78,13 @@ newtype ConfigsT m a = ConfigsT { unConfigsT :: ReaderT (Map Text ByteString) m 
     , MonadPlus
     , Alternative
     , MonadFix
+    , MonadThrow
     , MonadIO
     , MonadBase m'
     , MonadBaseControl m'
     , MonadRef
     , MonadTrans
+    , MFunctor
     , DomBuilder t
     , MonadHold t
     , MonadReflexCreateTrigger t